requirements.txt
-i https://pypi.org/simple
alembic==1.9.2; python_version >= '3.7'
aniso8601==9.0.1
appnope==0.1.3; sys_platform == 'darwin'
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1; python_version >= '3.6'
click==8.1.3; python_version >= '3.7'
decorator==5.1.1; python_version > '3.6' and python_version < '3.11'
executing==1.2.0
faker==16.6.0
flask==2.2.2
flask-bcrypt==1.0.1
flask-cors==3.0.10
flask-migrate==4.0.2
flask-restful==0.3.9
flask-sqlalchemy==3.0.2
greenlet==2.0.1; python_version >= '3' and (platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32'))))))
gunicorn==20.1.0
honcho==1.1.0
importlib-metadata==6.0.0; python_version < '3.10'
importlib-resources==5.10.2
ipdb==0.13.11
ipython==8.8.0; python_version > '3.6' and python_version < '3.11'
itsdangerous==2.1.2; python_version >= '3.7'
jedi==0.18.2; python_version >= '3.6'
jinja2==3.1.2; python_version >= '3.7'
jsonify==0.5
mako==1.2.4; python_version >= '3.7'
markupsafe==2.1.2; python_version >= '3.7'
matplotlib-inline==0.1.6; python_version >= '3.5'
parso==0.8.3; python_version >= '3.6'
pexpect==4.8.0; sys_platform != 'win32'
pickleshare==0.7.5
prompt-toolkit==3.0.36; python_full_version >= '3.6.2'
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pygments==2.14.0; python_version >= '3.6'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==0.21.0
pytz==2022.7.1
setuptools==66.0.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==1.4.46; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlalchemy-serializer==1.4.1
stack-data==0.6.2
tomli==2.0.1; python_version > '3.6' and python_version < '3.11'
traitlets==5.8.1; python_version >= '3.7'
wcwidth==0.2.6
werkzeug==2.2.2; python_version >= '3.7'
zipp==3.11.0; python_version < '3.10'
